sovits_svc/requirements.txt
fsspec
pyworld
matplotlib
soundfile
scikit-learn
scipy
tensorboard
transformers
tqdm
librosa
omegaconf
gradio==5.34.2
ruamel.yaml
resampy
numpy==1.24
chardet
faiss-cpu==1.7.4
gdown
wget
praat-parselmouth
